Change main scenario and add exception 4
Change the main success scenario. Now the ATC leases airspace to UAVs. The UAV holds the lease until the ATC leases it some replacement airspace.

Add exception 4. In case the ATC cannot grant a lease to the UAV, the ATC attaches a list of other reservations and a list of relevant airspace restrictions. The UAV uses these lists to decide on its next steps.